♻️ Refactor : 사진 스켈레톤
