+++ src/feature/picsel/shared/components/ui/atoms/Skeleton/GridPhotoSkeleton.tsx
