+++ src/feature/picsel/shared/components/ui/atoms/Skeleton/HorizontalPhotoSkeleton.tsx
