+++ src/feature/picsel/shared/components/ui/atoms/Skeleton/PhotoSkeleton.tsx
